move 'contrib' folder to '_plugins'
These are some inbuilt things so a contrib folder seems a bit misnamed
