+++ MANIFEST.in
@@ -1,4 +1,4 @@
 include tavern/schemas/tests.schema.yaml
-include tavern/contrib/mqtt/schema.yaml
-include tavern/contrib/rest/schema.yaml
+include tavern/_plugins/mqtt/schema.yaml
+include tavern/_plugins/rest/schema.yaml
 include LICENSE
