+++ setup.cfg
@@ -57,8 +57,8 @@ pytest11 =
     tavern = tavern.testutils.pytesthook
 
 tavern =
-    requests = tavern.contrib.rest.tavernhook:TavernRestPlugin
-    mqtt = tavern.contrib.mqtt.tavernhook
+    requests = tavern._plugins.rest.tavernhook:TavernRestPlugin
+    mqtt = tavern._plugins.mqtt.tavernhook
 
 
 [bdist_wheel]
