+++ tests/unit/test_mqtt.py
@@ -2,7 +2,7 @@ import pytest
 from mock import patch
 import paho.mqtt.client as paho
 
-from tavern.contrib.mqtt.client import MQTTClient
+from tavern._plugins.mqtt.client import MQTTClient
 from tavern.util import exceptions
