+++ tests/unit/test_request.py
@@ -8,7 +8,7 @@ except ImportError:
 
 import pytest
 
-from tavern.contrib.rest.request import RestRequest, get_request_args
+from tavern._plugins.rest.request import RestRequest, get_request_args
 from tavern.util import exceptions
